requirements.txt
lxml
requests
playwright
python-dotenv
yt-dlp